+++ pages/index.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import projects from '../data/projects'
 import Metatags from '../src/components/Metatags'
 import Page from '../src/components/page/Page'
@@ -10,28 +9,12 @@ interface HomeProps {
     projects: ProjectDataMinimal[]
 }
 
-const PAGE_SIZE = 12
-
 export default function Home({ projects }: HomeProps) {
-    const [page, setPage] = useState(0)
-
-    const hasNext = projects.length > (page + 1) * PAGE_SIZE
-
-    const loadNextPage = () => {
-        if (!hasNext) {
-            return
-        }
-
-        setPage((currentPage) => currentPage + 1)
-    }
-
     return (
-        <Page onScrollToBottom={loadNextPage}>
+        <Page>
             <Metatags />
-            <Spacer amount="5.5em" />
-            <ProjectsGrid
-                projects={projects.slice(0, (page + 1) * PAGE_SIZE)}
-            />
+            <Spacer amount="3.5em" />
+            <ProjectsGrid projects={projects} />
             <Spacer amount="3.5em" />
         </Page>
     )
