+++ styles/components/navbar/NavBar.module.sass
@@ -1,20 +1,14 @@
 .navbar
-    position: fixed
-
     background-color: var(--navbar-bg)
-
-    z-index: 2
 
     display: flex
     flex-direction: row
     justify-content: space-between
     align-items: center
 
-    width: 100%
-    height: 60px
+    // width: 100vw
+    min-height: 60px
     padding: 0 15px
-
-    box-sizing: border-box
 
 
     .links
