+++ styles/components/page/Page.module.sass
@@ -2,9 +2,6 @@
     display: flex
     flex-direction: column
     justify-content: flex-start
-    align-items: center
+    align-items: stretch
 
-    width: 100vw
-    height: 100vh
-
-    overflow: scroll
+    background-color: var(--bg)
