+++ styles/components/project-details/ProjectDetails.module.sass
@@ -1,124 +1,129 @@
-.container
-    padding: 100px 60px 40px 60px
+.detailsView
+    display: flex
+    flex-direction: column
+    align-items: center
 
-    width: 100%
-    max-width: 1000px
+    .container
+        padding: 60px
 
-    @media screen and ( max-width: 640px)
-        padding: 100px 0px
-
-    .previewImage
         width: 100%
-        filter: contrast(120%)
-        margin-bottom: 4em
-
-    .card
-        margin-bottom: 10px
-        padding: 15px 20px 30px
-        border-radius: 10px
-        background: var(--bg-content)
-
-    .header
-        display: flex
-        flex-direction: row
-        justify-content: space-between
-        align-items: flex-start
-        padding-top: 15px !important
-
-        padding: 20px 0
-
-        @media screen and ( max-width: 600px)
-            flex-direction: column
-            justify-content: flex-start
-
-            .gitHubLink
-                margin-top: 20px
-
-        @media screen and ( max-width: 1000px)
-            padding: 20px
-
-        .gitHubLink
-            max-width: 170px
-            min-width: 145px
-
-        .information
-            display: flex
-            flex-direction: column
-            justify-content: flex-start
-            align-items: flex-start
-
-            .title
-                display: flex
-                align-items: center
-
-                font-size: 1.8em
-                font-weight: 700
-                margin-bottom: 10px
-
-                .date
-                    margin-left: 10px
-                    margin-bottom: 0
-
-                .technology
-                    height: 40px
-                    width: 40px
-                    margin-right: 10px
-
-            .subtitle
-                font-size: 1.2em
-                font-weight: 500
-                font-style: italic
-                margin-bottom: 10px
-
-    .sectionBody
-        padding: 30px 0px 0px
-
-        @media screen and ( max-width: 1000px)
-            padding: 30px 20px 0px
-
-            &.mobileFullWidth
-                padding: 10px 0px
-
-                .sectionTitle
-                    padding: 0 20px
-
-        .sectionTitle
-            font-size: 1.3em
-            margin-bottom: 20px
-
-    .descriptionItems
-        .descriptionItem
-            font-size: 1em
-            line-height: 1.4em
-            font-weight: 500
-
-            margin-top: 8px
-
-            max-width: 900px
-
-    .implementationDetails
-        padding: 0 30px
-
-        .bullet
-            line-height: 1.5em
-
-    .externalLinks
-        display: flex
-        justify-content: flex-start
-        flex-wrap: wrap
-        width: 100%
-
-    .video
-        height: 600px
-        width: 100%
-        margin-bottom: 20px
+        max-width: 1000px
 
         @media screen and ( max-width: 640px)
-            height: 30vh
+            padding: 100px 0px
 
-        iframe
+        .previewImage
             width: 100%
-            height: 100%
+            filter: contrast(120%)
+            margin-bottom: 4em
 
-            // @media screen and ( min-width: 1000px)
-            //     border-radius: 15px
+        .card
+            margin-bottom: 10px
+            padding: 15px 20px 30px
+            border-radius: 10px
+            background: var(--bg-content)
+
+        .header
+            display: flex
+            flex-direction: row
+            justify-content: space-between
+            align-items: flex-start
+            padding-top: 15px !important
+
+            padding: 20px 0
+
+            @media screen and ( max-width: 600px)
+                flex-direction: column
+                justify-content: flex-start
+
+                .gitHubLink
+                    margin-top: 20px
+
+            @media screen and ( max-width: 1000px)
+                padding: 20px
+
+            .gitHubLink
+                max-width: 170px
+                min-width: 145px
+
+            .information
+                display: flex
+                flex-direction: column
+                justify-content: flex-start
+                align-items: flex-start
+
+                .title
+                    display: flex
+                    align-items: center
+
+                    font-size: 1.8em
+                    font-weight: 700
+                    margin-bottom: 10px
+
+                    .date
+                        margin-left: 10px
+                        margin-bottom: 0
+
+                    .technology
+                        height: 40px
+                        width: 40px
+                        margin-right: 10px
+
+                .subtitle
+                    font-size: 1.2em
+                    font-weight: 500
+                    font-style: italic
+                    margin-bottom: 10px
+
+        .sectionBody
+            padding: 30px 0px 0px
+
+            @media screen and ( max-width: 1000px)
+                padding: 30px 20px 0px
+
+                &.mobileFullWidth
+                    padding: 10px 0px
+
+                    .sectionTitle
+                        padding: 0 20px
+
+            .sectionTitle
+                font-size: 1.3em
+                margin-bottom: 20px
+
+        .descriptionItems
+            .descriptionItem
+                font-size: 1em
+                line-height: 1.4em
+                font-weight: 500
+
+                margin-top: 8px
+
+                max-width: 900px
+
+        .implementationDetails
+            padding: 0 30px
+
+            .bullet
+                line-height: 1.5em
+
+        .externalLinks
+            display: flex
+            justify-content: flex-start
+            flex-wrap: wrap
+            width: 100%
+
+        .video
+            height: 600px
+            width: 100%
+            margin-bottom: 20px
+
+            @media screen and ( max-width: 640px)
+                height: 30vh
+
+            iframe
+                width: 100%
+                height: 100%
+
+                // @media screen and ( min-width: 1000px)
+                //     border-radius: 15px
