+++ styles/global/index.sass
@@ -5,19 +5,15 @@ body
     font-family: 'Open Sans', sans-serif
 
     color: var(--text-color)
-    background-color: var(--bg)
+    background-color: black
 
     -webkit-font-smoothing: antialiased
     -moz-osx-font-smoothing: grayscale
     text-rendering: optimizeLegibility
 
-    height: 100vh
-    width: 100%
 
     button,
     input
         font-family: inherit
 
     #__next
-        height: 100%
-        width: 100%
